Conflicts:
	Makefile
	arch/arm/tools/mach-types
	drivers/input/misc/Kconfig
	drivers/mtd/nand/davinci_nand.c
	drivers/mtd/nand/nand_base.c
	drivers/regulator/Kconfig
	drivers/rtc/Makefile
	drivers/video/Makefile
	drivers/video/da8xx-fb.c
	drivers/watchdog/davinci_wdt.c

+++ drivers/watchdog/davinci_wdt.c
@@ -88,7 +88,7 @@ static void wdt_enable(void)
 {
 	u32 tgcr;
 	u32 timer_margin;
-	u32 wdt_freq;
+	unsigned long wdt_freq;
 
 	wdt_freq = clk_get_rate(wdt_clk);
